pyproject.toml
[tool.poetry]
name = "mailman-migrate"
version = "0.1.0"
description = ""
authors = ["Sébastiaan <se_bastiaan@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.7"
google-api-python-client = "^1.7.11"
requests = "^2.22.0"
google_auth_oauthlib = "^0.4.1"
tqdm = "^4.36.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
